AI 채팅 SSE 로 처리 (#77)
* feat: AI 채팅 API 요청 컨트롤러 구현

* chore: fix lint

* feat: service 에서 AI 서버 응답 가공해서 보내도록 수정, parsing 메서드들 ExternalApiUtil 로 분리

* chore: fix lint and warning

* chore: fix lint

---------
